rename User to StateAccessor
This will become a trait that can be backed by either a real client or a
set of test data. Long-term we might also want two client paths: one
that uses the sync endpoint to fetch all state at once and one that
makes a bunch of individual api calls. The sync endpoint approach is
pretty bad with current grapevine because we don't have filter support,
but might be better for synapse.