emit warnings when message history may be lost
We can't easily check for history lost due to a previous
m.room.history_visibility state when the current state is permissive. In
order to do that we would need to page through the entire history of
history visibility state in each room. In theory that might be doable
with filtered sync, but I suspect the performance would still be very
bad.

We should emit a warning about this limitation.