store plan errors in-band
This makes it easier to ask questions like "were there any errors for
this room id" which is needed for determining whether it is safe to
leave rooms on the old user. This refactor leaves the planning logic in
a suboptimal state. I tried to keep it as close as possible to the old
structure, and will refactor it further in a future commit.