throw an error when attempting to migrate between different servers
This doesn't seem to work with conduit derivatives because the join
needs to be sent after the invite is received over federation. There is
no efficient way to wait for invites to be received (sync is the only
option). Probably we just need to bite the bullet and do a sync loop,
but I'm gonna not worry about that for now.